requirements.txt
sentence-transformers
pinecone-client
langchain
langchain-community
flask
pypdf
python-dotenv
google-generativeai
